Update fork owner information
This change updates the fork owner for the repo. It seems that some of
the caching settings depend on the full repo name in order to work.
After making this change we were able to use the fork in a test PR on
our main application successfully.

Also: update CI workflows to check on pull request event rather than
pushes for the non-main branch.